requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
boto3>=1.26.0
aiohttp>=3.8.0
telethon>=1.24.0
pydub>=0.25.1